add QCustomPlot

+++ model.cpp
@@ -5,7 +5,7 @@ bool flagAbsOrInc=false;
 vector<double> absAngle(10,0);
 vector<double> incAngle(10,0);
 vector<int> absNum(10,0);//范围为-65536~65536
-vector<int> incNum(10,0);//采用int可测试电机持续转268圈便溢出，改为long
+vector<long> incNum(10,0);//采用int可测试电机持续转268圈便溢出，改为long
 DWORD devindex;//CAN设备索引
 DWORD devtype;//CAN设备类型
 UINT recIndex;
